[1.43.57] Indicate language changes in IPA phoneme output.
Phoneme definitions FMT(), second parameter is amplitude percentage.

Phoneme definitions CALL(),  If phoneme type is not defined, copy the static attributes from the called phoneme.

Unpronouncable word check: stop when an apostrophe is reached.


git-svn-id: https://espeak.svn.sourceforge.net/svnroot/espeak/trunk@250 d46cf337-b52f-0410-862d-fd96e6ae7743

+++ src/klatt.cpp
@@ -421,7 +421,7 @@ if(option_log_frames)
 #endif
 
 		out = resonator(&(kt_globals.rsn[Rout]),out);
-		temp = (out * wdata.amplitude * kt_globals.amp_gain0) ;   /* Convert back to integer */
+		temp = (int)(out * wdata.amplitude * kt_globals.amp_gain0) ;   /* Convert back to integer */
 
 
 		// mix with a recorded WAV if required for this phoneme
